+++ middleware/proxy/grpc_test.go
@@ -3,8 +3,6 @@ package proxy
 import (
 	"testing"
 	"time"
-
-	"google.golang.org/grpc/grpclog"
 )
 
 func pool() []*UpstreamHost {
@@ -19,8 +17,6 @@ func pool() []*UpstreamHost {
 }
 
 func TestStartupShutdown(t *testing.T) {
-	grpclog.SetLogger(discard{})
-
 	upstream := &staticUpstream{
 		from:        ".",
 		Hosts:       pool(),
@@ -56,13 +52,3 @@ func TestStartupShutdown(t *testing.T) {
 		t.Errorf("Shutdown didn't remove conns, found %d", len(g.conns))
 	}
 }
-
-// discard is a Logger that outputs nothing.
-type discard struct{}
-
-func (d discard) Fatal(args ...interface{})                 {}
-func (d discard) Fatalf(format string, args ...interface{}) {}
-func (d discard) Fatalln(args ...interface{})               {}
-func (d discard) Print(args ...interface{})                 {}
-func (d discard) Printf(format string, args ...interface{}) {}
-func (d discard) Println(args ...interface{})               {}
